requirements.txt
altgraph==0.17.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
ping3==4.0.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dateutil==2.9.0.post0
pytz==2024.1
pyasn1==0.5.0
pywin32-ctypes==0.2.3
requests==2.32.3
scapy==2.5.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
